--- project-moex/moex-app/src/main/java/ru/home/ilar/service/moex/FeignMoexIntegrationService.java
@@ -1,14 +1,0 @@
-package ru.home.ilar.service.moex;
-
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.GetMapping;
-import ru.home.ilar.dto.StockEntryDto;
-
-import java.util.List;
-
-@FeignClient(value = "moex-integration")
-public interface FeignMoexIntegrationService {
-
-    @GetMapping("/api/currentStockIndex")
-    List<StockEntryDto> getCurrentStockIndex();
-}
