+++ project-moex/moex-app/src/main/java/ru/home/ilar/service/moex/MoexIntegrationService.java
@@ -1,24 +1,14 @@
 package ru.home.ilar.service.moex;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import reactor.core.publisher.Mono;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.GetMapping;
 import ru.home.ilar.dto.StockEntryDto;
-import ru.home.ilar.service.StockIntegrationService;
 
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
 
-@Service
-@RequiredArgsConstructor
-public class MoexIntegrationService implements StockIntegrationService {
+@FeignClient(value = "moex-integration")
+public interface MoexIntegrationService {
 
-    private final FeignMoexIntegrationService moexIntegrationService;
-
-    @Override
-    public Mono<List<StockEntryDto>> getCurrentStockIndex() {
-        CompletableFuture<List<StockEntryDto>> future = CompletableFuture
-                .supplyAsync(moexIntegrationService::getCurrentStockIndex);
-        return Mono.fromFuture(future);
-    }
+    @GetMapping("/api/currentStockIndex")
+    List<StockEntryDto> getCurrentStockIndex();
 }
